feat: migrates assets
